Move functions to utils.py
This harmonizes imports in plotview.py and datadialogs.py
